+++ src/main/java/org/scion/jpan/ScionDatagramSocket.java
@@ -298,7 +298,8 @@ public class ScionDatagramSocket extends java.net.DatagramSocket {
           if (path == null) {
             path = channel.getPathPolicy().filter(channel.getOrCreateService2().getPaths(addr));
           } else if (path instanceof RequestPath
-              && ((RequestPath) path).getExpiration() > Instant.now().getEpochSecond()) {
+              && ((RequestPath) path).getMetadata().getExpiration()
+                  > Instant.now().getEpochSecond()) {
             // check expiration only for RequestPaths
             RequestPath request = (RequestPath) path;
             path = channel.getPathPolicy().filter(channel.getOrCreateService2().getPaths(request));
