+++ src/test/java/org/scion/jpan/api/DatagramSocketApiTest.java
@@ -618,8 +618,9 @@ class DatagramSocketApiTest {
           try {
             socket.send(packet);
             RequestPath newPath = socket.getConnectionPath();
-            assertTrue(newPath.getExpiration() > expiredPath.getExpiration());
-            assertTrue(Instant.now().getEpochSecond() < newPath.getExpiration());
+            assertTrue(
+                newPath.getMetadata().getExpiration() > expiredPath.getMetadata().getExpiration());
+            assertTrue(Instant.now().getEpochSecond() < newPath.getMetadata().getExpiration());
             // assertNull(channel.getCurrentPath());
           } catch (IOException e) {
             throw new RuntimeException(e);
@@ -660,7 +661,7 @@ class DatagramSocketApiTest {
             basePath.getRemoteAddress(),
             basePath.getRemotePort(),
             basePath.getFirstHopAddress());
-    assertTrue(Instant.now().getEpochSecond() > expiredPath.getExpiration());
+    assertTrue(Instant.now().getEpochSecond() > expiredPath.getMetadata().getExpiration());
     return expiredPath;
   }
