+++ src/test/java/org/scion/jpan/demo/ScmpEchoDemo.java
@@ -149,8 +149,8 @@ public class ScmpEchoDemo {
     // ").append(channel.getLocalAddress().getAddress().getHostAddress()).append(nl);
     sb.append("Using path:").append(nl);
     sb.append("  Hops: ").append(ScionUtil.toStringPath(path));
-    sb.append(" MTU: ").append(path.getMtu());
-    sb.append(" NextHop: ").append(path.getInterface().getAddress()).append(nl);
+    sb.append(" MTU: ").append(path.getMetadata().getMtu());
+    sb.append(" NextHop: ").append(path.getMetadata().getInterface().getAddress()).append(nl);
     println(sb.toString());
   }
